Use decimals instead of floats for prices

+++ src/Gir/Domain.fs
@@ -4,7 +4,7 @@ module Gir.Domain
 type Product =
     { ProductId: int
       Name: string
-      Price: float
+      Price: decimal
       Img: string }
 
 type CartEvent =
